windows-client/src/app/components/delete-dialog/delete-dialog.component.ts
import { Component, Inject } from '@angular/core';
import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
@Component({
    selector: 'app-delete-modal-dialog',
    templateUrl: './delete-dialog.component.html',
    styleUrls: ['./delete-dialog.component.scss'],
})
export class DeleteDialogComponent {
    constructor(private dialogRef: MatDialogRef<DeleteDialogComponent>, @Inject(MAT_DIALOG_DATA) public data: { action: string }) {}

    emitSupp(supp: boolean) {
        this.dialogRef.close(supp);
    }

    close() {
        this.dialogRef.close();
    }
}
